feat(native-world): complete generic network contract
Allow servers to select ordered format-3 sets of one through eight compatible packs while keeping aggregate admission, leases, LOD profiles, and registrar state derived from the selected identities.

Move the incompatible DM protocol to epoch 0x1DF, keep generic sets and extended low-precision Z within the effective 0x3D net-module ceiling, and advance closed startup authorization to 4/1/4. Decode retired 3/1/3 receipts only at the spent-ledger boundary so anti-replay history cannot block current publication or become activable.

Remove the 32000 invalid-model alias, reserve 65535 as the sentinel, place server model IDs above compact FileIDs, validate wide Lua/C++ inputs before narrowing, retain sufficient uint16 RPC fields, and fix negative extended-streamer cell alignment.

Validated with 158 extended-world tests (2 skipped), clean diff checks, Release Win32 Game SA/Client Core/Client Deathmatch builds, Release x64 Deathmatch build, a live three-pack server-selected set, city transitions and LOD fences, reconnect, resource restart, full server restart, and an actual 0x41DE client rejected by the 0x41DF server before join.

+++ Client/mods/deathmatch/logic/luadefs/CLuaBuildingDefs.h
@@ -19,7 +19,7 @@ public:
     static void AddClass(lua_State* luaVM);
 
     // Buiding create funcs
-    static CClientBuilding* CreateBuilding(lua_State* const luaVM, std::uint16_t modelId, CVector pos, std::optional<CVector> rot,
+    static CClientBuilding* CreateBuilding(lua_State* const luaVM, std::uint32_t modelId, CVector pos, std::optional<CVector> rot,
                                            std::optional<std::uint8_t> interior);
     static void             RemoveAllGameBuildings();
     static void             RestoreGameBuildings();
