+++ Server/mods/deathmatch/logic/luadefs/CLuaBuildingDefs.h
@@ -18,6 +18,6 @@ public:
     static void LoadFunctions();
     static void AddClass(lua_State* luaVM);
 
-    static CBuilding* CreateBuilding(lua_State* const luaVM, std::uint16_t modelId, CVector pos, std::optional<CVector> rot,
+    static CBuilding* CreateBuilding(lua_State* const luaVM, std::uint32_t modelId, CVector pos, std::optional<CVector> rot,
                                      std::optional<std::uint8_t> interior);
 };
